Replace `evented-select` with `ember-power-select`
Fixes #51
