+++ app/styles/app.scss
@@ -1,3 +1,4 @@
+@import 'ember-power-select';
 @import 'bourbon';
 @import 'reset';
 @import 'mixins/border-radius';
