+++ package.json
@@ -26,21 +26,22 @@
     "ember-cli-app-version": "^1.0.0",
     "ember-cli-babel": "^5.1.5",
     "ember-cli-content-security-policy": "0.5.0",
-    "ember-cli-sass": "^5.2.1",
-    "express": "^4.8.5",
-    "glob": "^4.0.5",
-    "http-proxy": "^1.8.1",
     "ember-cli-dependency-checker": "^1.2.0",
     "ember-cli-htmlbars": "^1.0.1",
     "ember-cli-htmlbars-inline-precompile": "^0.3.1",
     "ember-cli-inject-live-reload": "^1.3.1",
     "ember-cli-qunit": "^1.2.1",
     "ember-cli-release": "0.2.8",
+    "ember-cli-sass": "^5.2.1",
     "ember-cli-sri": "^2.0.0",
     "ember-cli-uglify": "^1.2.0",
     "ember-cli-yuidoc-converter": "git://github.com/trek/ember-cli-yuidoc-converter.git#7bc04b2",
     "ember-disable-proxy-controllers": "^1.0.1",
     "ember-export-application-global": "^1.0.4",
-    "ember-resolver": "^2.0.3"
+    "ember-power-select": "^0.8.1",
+    "ember-resolver": "^2.0.3",
+    "express": "^4.8.5",
+    "glob": "^4.0.5",
+    "http-proxy": "^1.8.1"
   }
 }
